requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.1
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==4.7.5
packaging==20.1
pep517==0.7.0
pkg-resources==0.0.0
progress==1.5
pyee==7.0.1
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
slackclient==2.5.0
slackeventsapi==2.1.0
urllib3==1.25.8
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==1.0.0
yarl==1.4.2
